Contact mail with event, queue added, whole menu finished

+++ database/migrations/2022_08_31_072718_create_food_items_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->timestamps();
             $table->json('title');
             $table->json('description');
-            $table->integer('price');
+            $table->decimal('price', 4, 2);
             $table->integer('order');
             $table->unsignedBigInteger('food_category_id')->index();
             $table->foreign('food_category_id')->references('id')->on('food_categories')->onDelete('cascade');
